Rag-pipeline/rag_service/requirements.txt
fastapi
uvicorn
pydantic

langchain
langchain-core
langchain-community
langchain-chroma
langchain-google-genai
langchain-text-splitters

chromadb
openai
python-dotenv
pypdf
requests
beautifulsoup4
